fix issue with gain tagging in Opus files (beets)
Beets automatically copies over metadata after converting the file,
which results in ReplayGain tags being added from FLAC sources.

This is bad because playing the RG tags on top of the track gain
results in too-quiet sound. opusenc applies this track gain
automatically, and ffmpeg does not; so switch back to ffmpeg.